+++ apps/docs/app/page.tsx
@@ -1,4 +1,4 @@
-import { Button } from "@viniciusvinnatest/ui/button";
+import { Button } from "@vinicius.vinna/ui/button";
 import Image from "next/image";
 
 export default function Home() {