+++ packages/foo/CHANGELOG.md
@@ -1,7 +1,7 @@
-# @viniciusvinnatest/foo
+# @vinicius.vinna/foo
 
 ## 0.0.1
 
 ### Patch Changes
 
-- [`d6d2719`](https://github.com/yrnana/turbo-monorepo-lib/commit/d6d2719a1fb0a081ebdd704382eed4c14bcd07eb) Thanks [@viniciusvinnatest](https://github.com/yrnana)! - Hello World
+- [`d6d2719`](https://github.com/yrnana/turbo-monorepo-lib/commit/d6d2719a1fb0a081ebdd704382eed4c14bcd07eb) Thanks [@vinicius.vinna](https://github.com/yrnana)! - Hello World